change: rearranged all screens
